Split off ScreenManager from Screen.cpp
