+++ UI/ChatScreen.cpp
@@ -6,6 +6,7 @@
 #include "Common/UI/View.h"
 #include "Common/UI/ViewGroup.h"
 #include "Common/UI/ScrollView.h"
+#include "Common/UI/ScreenManager.h"
 #include "Common/UI/UI.h"
 
 #include "Common/Data/Text/I18n.h"
