+++ UI/MainScreen.cpp
@@ -28,6 +28,7 @@
 #include "Common/UI/View.h"
 #include "Common/UI/ViewGroup.h"
 
+#include "Common/UI/ScreenManager.h"
 #include "Common/File/FileUtil.h"
 #include "Common/StringUtils.h"
 #include "Core/System.h"
