+++ UI/Store.cpp
@@ -22,6 +22,7 @@
 #include "Common/UI/ViewGroup.h"
 #include "Common/UI/IconCache.h"
 #include "Common/UI/ScrollView.h"
+#include "Common/UI/ScreenManager.h"
 #include "Common/Render/DrawBuffer.h"
 
 #include "Common/Log.h"
