+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -221,6 +221,7 @@
     <ClInclude Include="..\..\Common\UI\PopupScreens.h" />
     <ClInclude Include="..\..\Common\UI\Root.h" />
     <ClInclude Include="..\..\Common\UI\Screen.h" />
+    <ClInclude Include="..\..\Common\UI\ScreenManager.h" />
     <ClInclude Include="..\..\Common\UI\ScrollView.h" />
     <ClInclude Include="..\..\Common\UI\TabHolder.h" />
     <ClInclude Include="..\..\Common\UI\Tween.h" />
@@ -406,6 +407,7 @@
     <ClCompile Include="..\..\Common\UI\PopupScreens.cpp" />
     <ClCompile Include="..\..\Common\UI\Root.cpp" />
     <ClCompile Include="..\..\Common\UI\Screen.cpp" />
+    <ClCompile Include="..\..\Common\UI\ScreenManager.cpp" />
     <ClCompile Include="..\..\Common\UI\ScrollView.cpp" />
     <ClCompile Include="..\..\Common\UI\TabHolder.cpp" />
     <ClCompile Include="..\..\Common\UI\Tween.cpp" />
